% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rmse.R
\name{eval_rmse}
\alias{eval_rmse}
\title{Computes the Root Mean Square Error (RMSE)}
\usage{
eval_rmse(x, y, by.band = FALSE)
}
\arguments{
\item{x}{the high-resolution actual image as a \code{RasterStack}}

\item{y}{the high-resolution predicted image as a \code{RasterStack}}

\item{by.band}{if \code{FALSE}(default), gives the average RMSE of all bands}
}
\description{
Quantifies the RMSE between the bands of two images
}
\details{
\code{x} and \code{y} are assumed to have the same projection, spatial
resolution, and number of bands.
}
