% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fc.R
\name{fit_fc}
\alias{fit_fc}
\title{Applies the Fit-FC spatiotemporal data fusion model}
\usage{
fit_fc(f1, c1, c2, sngb.lr, sngb.wg, nsim, scale = c(0, 1))
}
\arguments{
\item{f1}{fine image from t1 as a \code{RasterBrick}}

\item{c1}{coarse image from t1 as a \code{RasterBrick}}

\item{c2}{coarse image from t2 as a \code{RasterBrick}}

\item{nsim}{number of similar pixels in each neighborhood}

\item{scale}{the dynamic range of the image (default, \code{c(0,255)}}

\item{r}{number that defines the radius of the neighborhood}

\item{spw}{number with the weight of the spatial dependence}
}
\value{
the fine image predicted at t2 as a \code{RasterBrick}
}
\description{
Predicts the fine image at a time t2 based on a fine image from a previous
moment time t1 and two coarse images, from t1 and t2
}
\details{
Applies the Fit-FC method as described in Wang and Atkinson (2018).
In the same study, authors set the Fit-FC model to \code{r=15} and 
\code{nsim=30} to downscale Sentinel-3 images (300x300m) to Sentinel-2
resolution (10x10m).
}
\references{
Wang, Q., & Atkinson, P. M. (2018). 
Spatio-temporal fusion for daily Sentinel-2 images.
Remote Sensing of Environment, 204, 31-42.
}
