\name{stfusion-package}
\alias{stfusion-package}
\alias{stfusion}
\docType{package}
\title{
\packageTitle{stfusion}
}
\description{
\packageDescription{stfusion}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stfusion}
\packageIndices{stfusion}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stfusion}

Maintainer: \packageMaintainer{stfusion}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
