% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvirt.R
\name{uvirt}
\alias{uvirt}
\title{Applies the UVIRT spatiotemporal data fusion model}
\usage{
uvirt(f.ts, c2, sngb.lr, sngb.wg, nsim, scale = c(0, 1))
}
\arguments{
\item{f.ts}{a series of fine images as a \code{RasterStack}}

\item{c2}{coarse image from t2 as a \code{RasterBrick}}

\item{sngb.lr}{number of neighboring pixels for the virtual image generation}

\item{sngb.wg}{number of neighboring pixels for the residual compensation}

\item{nsim}{number of similar pixels in each neighborhood}

\item{scale}{the dynamic range of the image (default, \code{c(0,1)}}
}
\value{
the fine image predicted at tk as a \code{RasterStack}
}
\description{
Predicts the fine image at a time tk based on a series of fine images around
tk and a coarse image from that time.
}
